[WFCORE-6815] Move Domain Mode stability test case outside of the Domain testsuite Jira issue: https://issues.redhat.com/browse/WFCORE-6815
